Don't allow to import an element with an already existing id.
